+++ libs/data/src/lib/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `a73ef2f2684e379d99148e15cef87b23aab95836`;
+export const extractsHash = `d0eda8bb2084c2edd3a3e9d3774f4cac8c3041dd`;