+++ libs/data/src/lib/lazy-files-list.ts
@@ -97,7 +97,7 @@ export const lazyFilesList = {
   },
   'desynth': {
     'fileName': 'desynth.json',
-    'hashedFileName': 'desynth.4d8d34df9d3d9f315a906a9e69bdca04c810f48b.json'
+    'hashedFileName': 'desynth.e9a4717eb744c37d4dfcb1d5ab7f247f76d1bb4d.json'
   },
   'diademTerritory': {
     'fileName': 'diadem-territory.json',
@@ -141,7 +141,7 @@ export const lazyFilesList = {
   },
   'fishingLogTrackerPageData': {
     'fileName': 'fishing-log-tracker-page-data.json',
-    'hashedFileName': 'fishing-log-tracker-page-data.dc06110f9163ecbce37604bf5dc79251e75c9721.json'
+    'hashedFileName': 'fishing-log-tracker-page-data.b5ca599a943fa2dbeb2fe36dadad4aa53bc7aded.json'
   },
   'fishingLog': {
     'fileName': 'fishing-log.json',
@@ -149,7 +149,7 @@ export const lazyFilesList = {
   },
   'fishingSources': {
     'fileName': 'fishing-sources.json',
-    'hashedFileName': 'fishing-sources.eee983befda2ee92e4aad49eeb76e31a5ff39f19.json'
+    'hashedFileName': 'fishing-sources.49c0420ae1647e969e90d5b8392b7088e45da13c.json'
   },
   'fishingSpots': {
     'fileName': 'fishing-spots.json',
@@ -165,7 +165,7 @@ export const lazyFilesList = {
   },
   'gardeningSources': {
     'fileName': 'gardening-sources.json',
-    'hashedFileName': 'gardening-sources.bdcbc8d973284ee00b05fa8a4fbedfb560fb5887.json'
+    'hashedFileName': 'gardening-sources.4fc8ce8355a5d30b6441634e343eee595e61b190.json'
   },
   'gatheringBonuses': {
     'fileName': 'gathering-bonuses.json',
@@ -197,7 +197,7 @@ export const lazyFilesList = {
   },
   'gcSupply': {
     'fileName': 'gc-supply.json',
-    'hashedFileName': 'gc-supply.e83fcd83e4ed02719926bd5646c2a31fbd0159ed.json'
+    'hashedFileName': 'gc-supply.76587febb3a65785771bab618583d0491259da50.json'
   },
   'gilShopNames': {
     'fileName': 'gil-shop-names.json',
@@ -229,7 +229,7 @@ export const lazyFilesList = {
   },
   'instanceSources': {
     'fileName': 'instance-sources.json',
-    'hashedFileName': 'instance-sources.f23cc445c051dbade24c9a3af60d82a81907334a.json'
+    'hashedFileName': 'instance-sources.b0371eaa854aaff131e2b5c50ec75dfa9222ab64.json'
   },
   'instances': {
     'fileName': 'instances.json',
@@ -345,7 +345,7 @@ export const lazyFilesList = {
   },
   'lootSources': {
     'fileName': 'loot-sources.json',
-    'hashedFileName': 'loot-sources.78f12ae7e566c298fd88101f165099bc76c489b4.json'
+    'hashedFileName': 'loot-sources.a77c76db82e30daebb155dad81765a70011f75fc.json'
   },
   'mapEntries': {
     'fileName': 'map-entries.json',
@@ -373,7 +373,7 @@ export const lazyFilesList = {
   },
   'mogstationSources': {
     'fileName': 'mogstation-sources.json',
-    'hashedFileName': 'mogstation-sources.c1711b630d66f2454e852ee2999cdf316933e310.json'
+    'hashedFileName': 'mogstation-sources.961b19b56b08c1918e47ff09b188b0324658c7ef.json'
   },
   'monsters': {
     'fileName': 'monsters.json',
@@ -445,7 +445,7 @@ export const lazyFilesList = {
   },
   'reverseInstanceSources': {
     'fileName': 'reverse-instance-sources.json',
-    'hashedFileName': 'reverse-instance-sources.3d5b170c4d452e71f21315182f5fcde5db6d59b8.json'
+    'hashedFileName': 'reverse-instance-sources.d0f495f87d63051cdf85c89fb18060d5f6cecb97.json'
   },
   'reverseReduction': {
     'fileName': 'reverse-reduction.json',
@@ -553,7 +553,7 @@ export const lazyFilesList = {
   },
   'voyageSources': {
     'fileName': 'voyage-sources.json',
-    'hashedFileName': 'voyage-sources.fa21d359cb701b71a4cc993bdff258249b0b6bb3.json'
+    'hashedFileName': 'voyage-sources.5e4593b471d213e31e72139596c3096106f347ba.json'
   },
   'weathers': {
     'fileName': 'weathers.json',
@@ -561,11 +561,11 @@ export const lazyFilesList = {
   },
   'koAchievementDescriptions': {
     'fileName': '/ko/ko-achievement-descriptions.json',
-    'hashedFileName': '/ko/ko-achievement-descriptions.311caf93a32fd7f33bed3514e7a15b4e68b99181.json'
+    'hashedFileName': '/ko/ko-achievement-descriptions.4dafdbe489f0108de5d0732a9eafaab0e0b5d58a.json'
   },
   'koAchievements': {
     'fileName': '/ko/ko-achievements.json',
-    'hashedFileName': '/ko/ko-achievements.edb6a5ac6c101ff5998ca3cce4dcecd5ef7e30c2.json'
+    'hashedFileName': '/ko/ko-achievements.5506e7067875f989e84874d451ad30a9fd983ed8.json'
   },
   'koActionCategories': {
     'fileName': '/ko/ko-action-categories.json',
@@ -573,11 +573,11 @@ export const lazyFilesList = {
   },
   'koActionDescriptions': {
     'fileName': '/ko/ko-action-descriptions.json',
-    'hashedFileName': '/ko/ko-action-descriptions.437b9b91e2c74716939cb1192b2ea7c4d1d523f6.json'
+    'hashedFileName': '/ko/ko-action-descriptions.21d17614d42093fc84a7d28a89501505befc7b1a.json'
   },
   'koActions': {
     'fileName': '/ko/ko-actions.json',
-    'hashedFileName': '/ko/ko-actions.eb43a4a366055a37d58c9e6743577fe2c9dd7f86.json'
+    'hashedFileName': '/ko/ko-actions.e0fabb309bd4f0db9b24a6b83dd23d205cb81cc5.json'
   },
   'koAirshipVoyages': {
     'fileName': '/ko/ko-airship-voyages.json',
@@ -605,7 +605,7 @@ export const lazyFilesList = {
   },
   'koEventItems': {
     'fileName': '/ko/ko-event-items.json',
-    'hashedFileName': '/ko/ko-event-items.4e269a00b96b7a1d7d7f30460e676edbfb590330.json'
+    'hashedFileName': '/ko/ko-event-items.70e016b92eeea43cdc694fc789953ef1e10e314f.json'
   },
   'koExVersions': {
     'fileName': '/ko/ko-ex-versions.json',
@@ -613,7 +613,7 @@ export const lazyFilesList = {
   },
   'koFates': {
     'fileName': '/ko/ko-fates.json',
-    'hashedFileName': '/ko/ko-fates.43a14666ded3cf8eaab81ffadef4bcb47a294732.json'
+    'hashedFileName': '/ko/ko-fates.f880150fa94f8b0a58d0f569cffc07bf6bfc3dfa.json'
   },
   'koFreeCompanyActions': {
     'fileName': '/ko/ko-free-company-actions.json',
@@ -621,7 +621,7 @@ export const lazyFilesList = {
   },
   'koGatheringBonuses': {
     'fileName': '/ko/ko-gathering-bonuses.json',
-    'hashedFileName': '/ko/ko-gathering-bonuses.dfefba1c5eb1d6197fdfa3cf8b5ff596f9e4faad.json'
+    'hashedFileName': '/ko/ko-gathering-bonuses.b1a18d5b43445abe501fcfabd50b64d0f9f79b97.json'
   },
   'koGatheringTypes': {
     'fileName': '/ko/ko-gathering-types.json',
@@ -629,11 +629,11 @@ export const lazyFilesList = {
   },
   'koInstanceDescriptions': {
     'fileName': '/ko/ko-instance-descriptions.json',
-    'hashedFileName': '/ko/ko-instance-descriptions.f7ae6f57de1b2afdf9cb3576e609ec233664f0f4.json'
+    'hashedFileName': '/ko/ko-instance-descriptions.4b5696a376c2bd2f913443ab853a65222d070164.json'
   },
   'koInstances': {
     'fileName': '/ko/ko-instances.json',
-    'hashedFileName': '/ko/ko-instances.6fc25eedc8d735568f2597e9d1bb2dc854a5aefd.json'
+    'hashedFileName': '/ko/ko-instances.f87c86fdb39c9981c21f6a67976288d03323ff98.json'
   },
   'koIslandCraftworksTheme': {
     'fileName': '/ko/ko-island-craftworks-theme.json',
@@ -641,7 +641,7 @@ export const lazyFilesList = {
   },
   'koItemDescriptions': {
     'fileName': '/ko/ko-item-descriptions.json',
-    'hashedFileName': '/ko/ko-item-descriptions.b372f1c124e31201898d065b0420616693441f6c.json'
+    'hashedFileName': '/ko/ko-item-descriptions.6f9b2e27b0a4ec85d19ff188b64113914c64ce47.json'
   },
   'koItemKinds': {
     'fileName': '/ko/ko-item-kinds.json',
@@ -657,7 +657,7 @@ export const lazyFilesList = {
   },
   'koItems': {
     'fileName': '/ko/ko-items.json',
-    'hashedFileName': '/ko/ko-items.ec6e7854d9110f4c547e30fa312a1d05cda44380.json'
+    'hashedFileName': '/ko/ko-items.f072ee8e5aa266960d7245667cf376038ae05a10.json'
   },
   'koJobAbbr': {
     'fileName': '/ko/ko-job-abbr.json',
@@ -673,11 +673,11 @@ export const lazyFilesList = {
   },
   'koJournalGenre': {
     'fileName': '/ko/ko-journal-genre.json',
-    'hashedFileName': '/ko/ko-journal-genre.2d484021e7eb35fd03d0242a6144d7b14086bfcb.json'
+    'hashedFileName': '/ko/ko-journal-genre.5eacae58c0e0754db43d524a180774226f25e0d0.json'
   },
   'koLeveDescriptions': {
     'fileName': '/ko/ko-leve-descriptions.json',
-    'hashedFileName': '/ko/ko-leve-descriptions.f6d13b3c38daa2f177c6f1f90091e9ec6a4874cd.json'
+    'hashedFileName': '/ko/ko-leve-descriptions.1a63c7f45af9cf92a81587f2671053bd6682f2cc.json'
   },
   'koLeves': {
     'fileName': '/ko/ko-leves.json',
@@ -685,43 +685,43 @@ export const lazyFilesList = {
   },
   'koMaps': {
     'fileName': '/ko/ko-maps.json',
-    'hashedFileName': '/ko/ko-maps.a68a97b3b4904466adc1678c69f0bf867ffaebc1.json'
+    'hashedFileName': '/ko/ko-maps.995bb3971687eaf2d292de2d50cffdc68b3e0eec.json'
   },
   'koMobs': {
     'fileName': '/ko/ko-mobs.json',
-    'hashedFileName': '/ko/ko-mobs.85e4125d2c9576ee0a0eaf2e62c6e040debcdf0e.json'
+    'hashedFileName': '/ko/ko-mobs.e75655a9e32ec316c8d026c9f1746ee01e32e850.json'
   },
   'koNotebookDivisionCategory': {
     'fileName': '/ko/ko-notebook-division-category.json',
-    'hashedFileName': '/ko/ko-notebook-division-category.ed0e3c45ca3f4637c87b61483651461261f14b4e.json'
+    'hashedFileName': '/ko/ko-notebook-division-category.3b406eca6cdab9ac29afde62be62a56d97acfa6d.json'
   },
   'koNotebookDivision': {
     'fileName': '/ko/ko-notebook-division.json',
-    'hashedFileName': '/ko/ko-notebook-division.4b07eadb28dc31d21934f474c7d60862c40829e9.json'
+    'hashedFileName': '/ko/ko-notebook-division.ba19dd9e9abf92d5c58d5ca02e49347ecb8c939e.json'
   },
   'koNpcTitles': {
     'fileName': '/ko/ko-npc-titles.json',
-    'hashedFileName': '/ko/ko-npc-titles.5d05e593fd796b40ddda2c12c7ad72a269a6a7cb.json'
+    'hashedFileName': '/ko/ko-npc-titles.a8c386eb1f5e00e6860188674cf0c62c34397681.json'
   },
   'koNpcs': {
     'fileName': '/ko/ko-npcs.json',
-    'hashedFileName': '/ko/ko-npcs.1cab1eb6c0b3c18e8207acbdddb815ba6283010e.json'
+    'hashedFileName': '/ko/ko-npcs.61dd1a0bce478ab41e80d614059d8fce02b5ee02.json'
   },
   'koPatchs': {
     'fileName': '/ko/ko-patchs.json',
-    'hashedFileName': '/ko/ko-patchs.3d4c17f36984cfca6f23d390d7d1564dbc5e3c57.json'
+    'hashedFileName': '/ko/ko-patchs.7dac644f39dbda999a32cf969ca15ceb06580c00.json'
   },
   'koPlaces': {
     'fileName': '/ko/ko-places.json',
-    'hashedFileName': '/ko/ko-places.f16306c0a663d1e61ba0113d790b378d14300436.json'
+    'hashedFileName': '/ko/ko-places.1245ec6466e7037f34027b3bb3b6a908579b3349.json'
   },
   'koQuestDescriptions': {
     'fileName': '/ko/ko-quest-descriptions.json',
-    'hashedFileName': '/ko/ko-quest-descriptions.4021fe2666ce1b8e9012feca17331f824c75a36a.json'
+    'hashedFileName': '/ko/ko-quest-descriptions.26f4cd4b53f5f95a92a8706146d605b0ec3b06ab.json'
   },
   'koQuests': {
     'fileName': '/ko/ko-quests.json',
-    'hashedFileName': '/ko/ko-quests.622d46b59114c55ec5fc858232e1532e7aefa3fe.json'
+    'hashedFileName': '/ko/ko-quests.61856e506553d38456cef6aa018742e8b665de55.json'
   },
   'koRaces': {
     'fileName': '/ko/ko-races.json',
@@ -729,27 +729,27 @@ export const lazyFilesList = {
   },
   'koRecipes': {
     'fileName': '/ko/ko-recipes.json',
-    'hashedFileName': '/ko/ko-recipes.f5d3a36f568121ef584f2c05e0a6f7961693f343.json'
+    'hashedFileName': '/ko/ko-recipes.89422f3396194eec5771c22278552972f411d0f5.json'
   },
   'koShops': {
     'fileName': '/ko/ko-shops.json',
-    'hashedFileName': '/ko/ko-shops.c24ed03938a9bbf512c4140fee6527ce3d2f9354.json'
+    'hashedFileName': '/ko/ko-shops.df8cdae94bc493157315ce9ea4f7e54a55142b4a.json'
   },
   'koStatusDescriptions': {
     'fileName': '/ko/ko-status-descriptions.json',
-    'hashedFileName': '/ko/ko-status-descriptions.ce69b3c0284edb88db1ed0d3c472505be4b9c19d.json'
+    'hashedFileName': '/ko/ko-status-descriptions.becfde626fd7279a308ea81f8c5671b57112e64e.json'
   },
   'koStatuses': {
     'fileName': '/ko/ko-statuses.json',
-    'hashedFileName': '/ko/ko-statuses.7d40bf90f8f81fda03b633dd5440b16a8bf57b27.json'
+    'hashedFileName': '/ko/ko-statuses.876bf069c67c30fb052dca8a2ba856a3f5db3432.json'
   },
   'koSubmarineVoyages': {
     'fileName': '/ko/ko-submarine-voyages.json',
-    'hashedFileName': '/ko/ko-submarine-voyages.c1eba6efed62357f49ac7af414165287b00d8c15.json'
+    'hashedFileName': '/ko/ko-submarine-voyages.5e8749d5b65eebcdba62324bf9d87ec8b42de76b.json'
   },
   'koTraitDescriptions': {
     'fileName': '/ko/ko-trait-descriptions.json',
-    'hashedFileName': '/ko/ko-trait-descriptions.0b67325908365c03a68e2f372deee852d0a4195e.json'
+    'hashedFileName': '/ko/ko-trait-descriptions.c96fe402d26562ec9f16a9c31d1d0400b6b3df60.json'
   },
   'koTraits': {
     'fileName': '/ko/ko-traits.json',
@@ -769,7 +769,7 @@ export const lazyFilesList = {
   },
   'koVentures': {
     'fileName': '/ko/ko-ventures.json',
-    'hashedFileName': '/ko/ko-ventures.dc88b821a6edc9d36dbd23186d22a7ccc2170990.json'
+    'hashedFileName': '/ko/ko-ventures.7bea0f3787df5be6ba08b1d13c6354198088d979.json'
   },
   'koWeathers': {
     'fileName': '/ko/ko-weathers.json',